ScrapperMLSchedule/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.4.2

# AI and Browser automation
ollama==0.1.0
aiohttp==3.9.1
websockets==12.0

# Testing
pytest==7.4.0
pytest-asyncio==0.23.0
httpx>=0.25.2,<0.26.0

# Type checking and linting
mypy==1.8.0
black==23.12.1
isort==5.13.2

# Logging and monitoring
python-json-logger==2.0.7

# Development tools
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
